feat: add unstyled reply comment
